Сделано

+++ Modul8Task1/Program.cs
@@ -11,25 +11,28 @@ namespace Modul8Task1
             string folder = @"D:\Temp";
             DateTime oldDate = DateTime.Now.AddMinutes(-30);
 
-            Console.WriteLine("old DateTime = "+oldDate);
-            PrintOldFiles(folder, oldDate);
+            // Ввод и провека папки на существование
+            Console.WriteLine("Введите директорию для зачистки :");
+            folder = Console.ReadLine();
+
+            if ( !Directory.Exists(folder))
+            {
+                Console.WriteLine(" Вы ввели не сущестdующую директорию. Программа прекращает свою работу.");
+                Console.ReadKey();
+                return;
+            }
+
+            Console.WriteLine("Начата зачистка директории {0} от файлов и директорий старше {1} ", folder, oldDate);
+
+            DeleteFolder(folder, oldDate);
+
+            Console.WriteLine("Зачистка завершена");
 
             Console.ReadKey();
         }
 
-        static void PrintOldFiles(string aFolder, DateTime aOldDate)  //  Вспомогательное для отработки лямбда по времени обращения
-        {
-            DirectoryInfo dirInfo = new DirectoryInfo(aFolder);
-            FileInfo[] arrFile = dirInfo.GetFiles();
-            
-            foreach (FileInfo file in arrFile.Where( d => d.LastAccessTime < aOldDate))
-            {
-                Console.WriteLine(file.LastAccessTime);
-            }
 
-        }
-
-        static void DeleteFolder(string aFolder)
+        static void DeleteFolder(string aFolder, DateTime aoldDate)   // Зачистка директории от устаревшей информации
         {
             try
             {
@@ -37,15 +40,15 @@ namespace Modul8Task1
                 DirectoryInfo[] arrDir = dirInfo.GetDirectories();
                 FileInfo[] arrFile = dirInfo.GetFiles();
 
-                foreach (FileInfo file in arrFile)
+                foreach (FileInfo file in arrFile.Where(d => d.LastAccessTime < aoldDate))
                 {
                     file.Delete();
                 }
 
                 foreach (DirectoryInfo dir in arrDir)
                 {
-                    DeleteFolder(dir.FullName);
-                    if (dir.GetDirectories().Length == 0 && dir.GetFiles().Length == 0)
+                    DeleteFolder(dir.FullName, aoldDate);
+                    if (dir.GetDirectories().Length == 0 && dir.GetFiles().Length == 0 && dir.LastAccessTime < aoldDate )
                     {
                         dir.Delete();
                     }
@@ -53,8 +56,20 @@ namespace Modul8Task1
             }
             catch (Exception ex)
             {
-                Console.WriteLine("Ну удалось зачистить устаревшую информаци по причине " + ex.Message);
+                Console.WriteLine("Не удалось зачистить устаревшую информаци по причине " + ex.Message);
             }
+        }
+
+        static void PrintOldFiles(string aFolder, DateTime aOldDate)  //  Вспомогательное для отработки лямбда по времени обращения
+        {
+            DirectoryInfo dirInfo = new DirectoryInfo(aFolder);
+            FileInfo[] arrFile = dirInfo.GetFiles();
+
+            foreach (FileInfo file in arrFile.Where(d => d.LastAccessTime < aOldDate))
+            {
+                Console.WriteLine(file.FullName +"    "+ file.LastAccessTime);
+            }
+
         }
 
     }
